pyproject.toml
[tool.poetry]
name = "fabrique-studio-test"
version = "0.1.0"
description = ""
authors = ["Nikita Ustiuzhanin <zaebumba@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.5"
djangorestframework = "^3.14.0"
django-timezone-field = "^6.0.1"
celery = "^5.3.4"
redis = "^5.0.1"
django-celery-beat = "^2.5.0"
psycopg2 = "^2.9.7"
requests = "^2.31.0"
drf-yasg = "^1.21.7"
django-prometheus = "^2.3.1"
gunicorn = "^21.2.0"


[tool.poetry.group.dev.dependencies]
isort = "^5.12.0"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
